+++ src/bkow/meow/processing/linear/predicate/error/visualization/sketch/Determinate.java
@@ -7,24 +7,24 @@ public class Determinate {
     }
 	  
     public static int DET4(Point p1, Point p2, Point p3, Point tp){
-	  float p1x = p1.getX();
-	  float p1y = p1.getY();
-	  float p2x = p2.getX();
-	  float p2y = p2.getY();
-	  float p3x = p3.getX();
-	  float p3y = p3.getY();
-	  float tpx = tp.getX();
-	  float tpy = tp.getY();
+	  double p1x = p1.getX();
+	  double p1y = p1.getY();
+	  double p2x = p2.getX();
+	  double p2y = p2.getY();
+	  double p3x = p3.getX();
+	  double p3y = p3.getY();
+	  double tpx = tp.getX();
+	  double tpy = tp.getY();
 	  
-	  float p1s = p1x*p1x + p1y*p1y;
-	  float p2s = p2x*p2x + p2y*p2y;
-	  float p3s = p3x*p3x + p3y*p3y;
-	  float tps = tpx*tpx + tpy*tpy;
+	  double p1s = p1x*p1x + p1y*p1y;
+	  double p2s = p2x*p2x + p2y*p2y;
+	  double p3s = p3x*p3x + p3y*p3y;
+	  double tps = tpx*tpx + tpy*tpy;
 	  return sign( p1x*p2y*p3s  +  p2x*p3y*tps  +  p3x*tpy*p1s  +  tpx*p1y*p2s
 	             -(p1s*p2y*p3x) - (p2s*p3y*tpx) - (p3s*tpy*p1x) - (tps*p1y*p2x) );
     }
     
-    public static int sign(float num){
+    public static int sign(double num){
 	  if (num > 0)
 	  {
 	    return 1;
